# server/openapi/hisab-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hisab API",
    "version": "1.0.0-alpha.1",
    "description": "واجهة حساب الموثقة للمؤسسات والفواتير والمزامنة المحلية أولاً.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/usingn50/hisab-app/blob/main/LICENSE"
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/v1",
      "description": "واجهة الإصدار الأول"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "رمز وصول قصير العمر صادر عن خدمة حساب."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "$ref": "#/components/schemas/RequestId"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "field",
                    "reason"
                  ]
                },
                "default": []
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RequestId": {
        "type": "string",
        "format": "uuid",
        "description": "معرف طلب فريد للتتبع."
      },
...
      "AuthSession": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "phone": {
                "$ref": "#/components/schemas/PhoneE164"
              },
              "displayName": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "id",
              "phone",
              "displayName"
            ]
          },
          "organizations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "baseCurrencyCode": {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 3
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "owner",
                    "manager",
                    "cashier",
                    "accountant",
                    "viewer"
                  ]
                }
              },
              "required": [
                "id",
                "name",
                "baseCurrencyCode",
                "role"
              ]
            }
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresAt",
          "user",
          "organizations"
        ]
      },
      "PhoneE164": {
        "type": "string",
        "pattern": "^\\+[1-9]\\d{7,14}$",
        "example": "+967712345678"
      },
      "OtpRequestResponse": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "expiresInSeconds": {
            "type": "integer",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "accepted",
          "expiresInSeconds"
        ]
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/OrganizationId"
          },
          "name": {
            "type": "string"
          },
          "baseCurrencyCode": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "timezone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "manager",
              "cashier",
              "accountant",
              "viewer"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "baseCurrencyCode",
          "timezone",
          "role"
        ]
      },
      "OrganizationId": {
        "type": "string",
        "format": "uuid"
      },
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": [
              "string",
              "null"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "isActive"
        ]
      },
      "OrganizationMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "manager",
              "cashier",
              "accountant",
              "viewer"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "userId",
          "phone",
          "role",
          "isActive"
        ]
      },
...
      "DeviceId": {
        "type": "string",
        "format": "uuid"
      }
    },
...
  },
  "paths": {
    "/auth/otp/request": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "requestOtp",
        "summary": "طلب رمز تحقق للهاتف",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "$ref": "#/components/schemas/PhoneE164"
                  }
                },
                "required": [
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم قبول طلب الرمز.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "رقم الهاتف غير صالح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "تجاوز المستخدم حد الطلبات.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "verifyOtp",
        "summary": "التحقق من رمز الهاتف وإنشاء جلسة",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "$ref": "#/components/schemas/PhoneE164"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  },
                  "deviceId": {
                    "$ref": "#/components/schemas/DeviceId"
                  }
                },
                "required": [
                  "phone",
                  "code",
                  "deviceId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم التحقق وأنشئت الجلسة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "401": {
            "description": "الرمز غير صالح أو منتهي.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "بيانات الطلب غير صالحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "refreshSession",
        "summary": "تجديد جلسة الجهاز",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 40
                  },
                  "deviceId": {
                    "$ref": "#/components/schemas/DeviceId"
                  }
                },
                "required": [
                  "refreshToken",
                  "deviceId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تدوير رموز الجلسة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "401": {
            "description": "جلسة التجديد غير صالحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "logoutSession",
        "summary": "إبطال جلسة الجهاز",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 40
                  },
                  "deviceId": {
                    "$ref": "#/components/schemas/DeviceId"
                  }
                },
                "required": [
                  "refreshToken",
                  "deviceId"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "أبطلت الجلسة."
          },
          "422": {
            "description": "بيانات الطلب غير صالحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "operationId": "listOrganizations",
        "summary": "قراءة المؤسسات التي ينتمي إليها المستخدم",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "المؤسسات المتاحة للمستخدم.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          },
          "401": {
            "description": "رمز الوصول غير صالح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "operationId": "createOrganization",
        "summary": "إنشاء مؤسسة مع فرعها الرئيسي",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "baseCurrencyCode": {
                    "type": "string",
                    "enum": [
                      "YER",
                      "USD",
                      "SAR"
                    ],
                    "default": "YER"
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "default": "Asia/Aden"
                  },
                  "initialBranchName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120,
                    "default": "الفرع الرئيسي"
                  },
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "أنشئت المؤسسة وعضوية المالك والفرع الرئيسي.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "401": {
            "description": "رمز الوصول غير صالح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "بيانات المؤسسة غير صالحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organizationId}/branches": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "operationId": "listBranches",
        "summary": "قراءة فروع مؤسسة مسموح بها للمستخدم",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrganizationId"
            },
            "required": true,
            "name": "organizationId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "فروع المؤسسة.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "401": {
            "description": "رمز الوصول غير صالح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "المستخدم لا ينتمي إلى المؤسسة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organizationId}/members": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "operationId": "addOrganizationMember",
        "summary": "إضافة عضو مسجل إلى مؤسسة",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrganizationId"
            },
            "required": true,
            "name": "organizationId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{7,14}$"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "manager",
                      "cashier",
                      "accountant",
                      "viewer"
                    ]
                  }
                },
                "required": [
                  "phone",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "أضيف العضو أو أعيد تفعيل عضويته.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMember"
                }
              }
            }
          },
          "401": {
            "description": "رمز الوصول غير صالح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "الدور الحالي لا يسمح بإدارة الأعضاء.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "المستخدم غير مسجل بعد أو العضوية موجودة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "بيانات العضو غير صالحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}